added historic for mouvement stock

+++ src/main/java/com/tricol/CommandeFournisseur/model/dto/ProduitCommandeDto.java
@@ -8,5 +8,5 @@ import lombok.*;
 @Builder
 public class ProduitCommandeDto {
     private Integer produitId;
-    private Double quantite;
+    private Integer quantite;
 }
